refactor: hippod dir to root dir
